# src/budget-fe/src/gc-client/models/BudgetedAccountResponse.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type BudgetedAccountResponse = {
    budgetedAccountId?: number;
    accountId?: string | null;
    fullName?: string | null;
};