# src/budget-fe/src/gc-client/models/GetBudgetsInRangeResponse.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { BudgetResponse } from './BudgetResponse';

export type GetBudgetsInRangeResponse = {
    budgets?: Array<BudgetResponse> | null;
};