# src/budget-fe/src/gc-client/models/ListBudgetedAccountsResponse.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { BudgetedAccountResponse } from './BudgetedAccountResponse';

export type ListBudgetedAccountsResponse = {
    accounts?: Array<BudgetedAccountResponse> | null;
};