# src/budget-fe/src/gc-client/models/SetBudgetRequest.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type SetBudgetRequest = {
    budgetId?: number;
    amount?: number;
    accountId?: number;
    date?: string;
};