# src/budget-fe/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "GnuCashBudget.Api",
    "version": "1.0"
  },
  "paths": {
    "/accounts": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Account"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Account"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Account"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/budgeted-accounts": {
      "get": {
        "tags": [
          "BudgetedAccounts"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ListBudgetedAccountsResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListBudgetedAccountsResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListBudgetedAccountsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "BudgetedAccounts"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddBudgetedAccountRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddBudgetedAccountRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddBudgetedAccountRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/budgets": {
      "put": {
        "tags": [
          "Budgets"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetBudgetRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetBudgetRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SetBudgetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Budgets"
        ],
        "parameters": [
          {
            "name": "fromDate",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetBudgetsInRangeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetBudgetsInRangeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetBudgetsInRangeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/expense-accounts": {
      "get": {
        "tags": [
          "ExpenseAccounts"
        ],
        "parameters": [
          {
            "name": "bottomLevelOnly",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Account"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Account"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Account"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Account": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "commodity": {
            "type": "string",
            "nullable": true
          },
          "accountType": {
            "$ref": "#/components/schemas/AccountType"
          },
          "childAccounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Account"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AccountType": {
        "enum": [
          "root",
          "asset",
          "stock",
          "mutual",
          "bank",
          "cash",
          "mutualFund",
          "otherAssets",
          "income",
          "expense",
          "liability",
          "credit",
          "equity"
        ],
        "type": "string"
      },
      "AddBudgetedAccountRequest": {
        "type": "object",
        "properties": {
          "accountGuid": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BudgetResponse": {
        "type": "object",
        "properties": {
          "budgetId": {
            "type": "integer",
            "format": "int32"
          },
          "budgetedAccountId": {
            "type": "integer",
            "format": "int32"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "BudgetedAccountResponse": {
        "type": "object",
        "properties": {
          "budgetedAccountId": {
            "type": "integer",
            "format": "int32"
          },
          "accountId": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetBudgetsInRangeResponse": {
        "type": "object",
        "properties": {
          "budgets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BudgetResponse"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ListBudgetedAccountsResponse": {
        "type": "object",
        "properties": {
          "accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BudgetedAccountResponse"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SetBudgetRequest": {
        "type": "object",
        "properties": {
          "budgetId": {
            "type": "integer",
            "format": "int32"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "accountId": {
            "type": "integer",
            "format": "int32"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      }
    }
  }
}